Fix pandocPipeTable highlighting
For some of my Markdown projects, I use kramdown which has support
for block inline attribute lists. The code to check for piped tables
doesn't quite play nicely when I try to add a class to it. For example,
in the following snippit the table never ends according to the current
code:

```markdown
| Heading 1         | Heading 2 |
| ---------         | --------- |
| vim-pandoc-syntax | is        |
| so                | great     |
{:sparkling}

## Not Styled as Heading!

Oh noes :'(
```

This fixes the problem by just checking to see if there's a `{`
character there.

# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
#
# Date:      Fri Sep 25 02:38:19 2015 -0400
#
# On branch fix/pandoc-pipe-tables
# Changes to be committed:
#	modified:   syntax/pandoc.vim
#