Fix automatic links by restricting characters in mail links.
Previously, pandocAutomaticLink captures too much in the e-mail side. For example, the following is captured as a single e-mail link:

    <div>Here is a sentence with a @citation. Here is another sentence.</div>

The presence of the @ and . in between any < and >, regardless of other content will result in a match.

What's needed is to restrict e-mail addresses to a certain range of characters, not including the space character. (See https://en.wikipedia.org/wiki/Email_address#Syntax.) This patch fixes this.