Correctly highlight links without suffix (#380)
This is a revert of ad853fe0. The original changes in ad853fe0 claimed
to be in service of this snippet:

    Here is some mono space linked text: [`project`](https://github.com/vim-pandoc/vim-pandoc-syntax).
    If I put here more `mono spaced` text, afterwards, the rest of the document is highlited wrong, try it yourself.

This snippet highlights completely fine, even without the changes made
in ad853fe0.

On the other hand, this change prevents this snippet from highlighting
the `[anchor]` inside the sentence:

    This is an [anchor] tag that should be auto-linked.

    [anchor]: https://example.com

I would love to revert the original change. Reverting it seems purely
additive to me (does not break old behaviors, adds new behaviors).